Support all operands in the `replace` operator
